requirements.txt
# FastAPI School Management System Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
bcrypt==4.0.1
slowapi==0.1.5

python-multipart==0.0.8
